rename IntroScreen container + view to GameSelect
